+++ src/main/java/com/yunche/loan/web/controller/PaddingCompanyController.java
@@ -43,13 +43,13 @@ public class PaddingCompanyController {
     }
 
     @GetMapping(value = "/delete")
-    public ResultBean<Void> delete(@RequestParam("id") Integer id) {
+    public ResultBean<Void> delete(@RequestParam("id") Long id) {
         logger.info(Arrays.asList("delete", id.toString()).stream().collect(Collectors.joining("-")));
         return paddingCompanyService.delete(id);
     }
 
     @GetMapping("/getById")
-    public ResultBean<PaddingCompanyVO> getById(@RequestParam("id") Integer id) {
+    public ResultBean<PaddingCompanyVO> getById(@RequestParam("id") Long id) {
         logger.info(Arrays.asList("getById", id.toString()).stream().collect(Collectors.joining("-")));
         return paddingCompanyService.getById(id);
     }
